pyproject.toml
[project]
name = "git-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "SillyDog5254", email = "ruyileng26@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.9.2",
    "gitpython>=3.1.44",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]

[project.scripts]
git-mcp-server = "app.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
